test only case sensitivity on darwin
