+++ api/iteration_test.go
@@ -8,37 +8,6 @@ import (
 	"unicode/utf8"
 )
 
-func TestNewIteration_CaseSensitive(t *testing.T) {
-	_, path, _, _ := runtime.Caller(0)
-	dir := filepath.Join(path, "..", "..", "fixtures", "iteration")
-
-	testCases := []map[string][]string{
-		{
-			"file": []string{filepath.Join(dir, "python", "leap", "one.py")},
-		},
-		{
-			"file": []string{filepath.Join(dir, "Python", "leap", "one.py")},
-		},
-		{
-			"file": []string{filepath.Join(dir, "Python", "Leap", "one.py")},
-		},
-	}
-
-	for _, testCase := range testCases {
-		iter, err := NewIteration(dir, testCase["file"])
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		if iter.TrackID != "python" {
-			t.Errorf("Expected language to be python, was %s", iter.TrackID)
-		}
-		if iter.Problem != "leap" {
-			t.Errorf("Expected problem to be leap, was %s", iter.Problem)
-		}
-	}
-}
-
 func TestNewIteration(t *testing.T) {
 	_, path, _, _ := runtime.Caller(0)
 	dir := filepath.Join(path, "..", "..", "fixtures", "iteration")
